Create EvenOdd.java
Return an array that contains the exact same numbers as the given array, but rearranged so that all the even numbers come before all the odd numbers.You can make a new array and return it.Note:Retain the order of other elements.
Sample Input: 1. evenOdd({1, 0, 1, 0, 0, 1, 1}) -> {0, 0, 0, 1, 1, 1, 1} 2. evenOdd({3, 3, 2})->{2, 3, 3}
Sample Output: 3. evenOdd({2, 2, 2}) -> {2, 2, 2}