Create post4.java
Given a non-empty array of ints, return a new array containing the elements from the original array that come after the last 4 in the original array. The original array will contain at least one 4. If the 4 is the last element in the array return the Original Arry.
Sample Input: 1. post4({2, 4, 1, 2}) -> {1, 2} 2. post4({4, 1, 4, 2}) -> {2} 3. post4({4, 4, 1, 2, 3}) -> {1, 2, 3}
Sample Output: post4({4, 1, 2, 3,4}) -> {4, 1, 2, 3,4}