Create isEverywhere.java
We'll say that a value is "everywhere" in an array if for every pair of adjacent elements in the array, at least one of the pair is that value. Return true if the given value is everywhere in the array.
Sample Input: 1. isEverywhere({1, 2, 1, 3}, 1) -> true 2. isEverywhere({1, 2, 1, 3}, 2) -> false
Sample Output: 3.isEverywhere({1, 2, 1, 3, 4}, 1)-> false